Podcasts updates.
Please can you remove the BACP logo on the home page.
On the FAQ about working online, please can we change those circles to either the question mark or the numbers on the home page.
On FAQ please can we make WHAT IS PSYCHOTHERAPY? Bold.
On Blog can we change opinion piece to BLOG ARTICLE.
I like the question mark thing, can we try a small one to the left of Nujoji Calvocoressi Psychotherapy in the header?
Can you give me a bit more chin in my profile pic pls.
Is there a way that when you get to the bottom of the page it scrolls onto the next page?
Can we move the two new books to the top of the book section (Klein and Hirsch).
Let’s move down EMDR and ARE YOU LOST IN THE WORLD LIKE ME to the bottom videos and replace thumbnail with the rat attached.
Kirk Honda should be Dr Kirk Honda
Please can you replace the room image with ROOM 4 Geo.
Can we make the dummy a little darker like in the attached.
On the contact page can we hyperlink the addresses in the right box and change the map square on the left to a big question mark and hyperlink to https://en.wikipedia.org/wiki/Question_mark and maybe a small thumbnail photo of me in the corner, perhaps in black and white. This is a different image.
Do you think we should display an email address on the website if people don’t want to use the box?
Is something happening to Google Analytics? Is it closing down?
On the Inside Time article, when you expand, please can it show the Inside Time logo.
What do you think that on each page somewhere different on the squares section on the lines we have a small butterfly? I might be getting carried away but could look interesting. Maybe a little bigger than the ones on the chairs.
Do I need a pop-up cookie banner?
Can we try these background colours (attached) going dark – light in the order of the tabs, left – right.
On the home page – underneath bottom right-hand square in the white text box can we do a CTA to contact and remove the one from the top box.
On the blog page top right can we do a CTA to FAQ.
https://kijo.co.uk/blog/luxury-web-design-techniques-2021/ I like the custom curser.
Can we do some hidden animation on the navigation buttons? Make them jiggle or something.